+++ src/helpers/constants.js
@@ -1,0 +1,10 @@
+const Position = {
+  BEFORE_BEGIN: `beforebegin`,
+  AFTER_BEGIN: `afterbegin`,
+  BEFORE_END: `beforeend`,
+  AFTER_END: `afterend`
+};
+
+const POINT_COUNT = 3;
+
+export {Position, POINT_COUNT};
