+++ src/helpers/utils.js
@@ -1,0 +1,5 @@
+const render = (container, element, place) => {
+  container.insertAdjacentHTML(place, element);
+};
+
+export {render};
