+++ src/view/day-list.js
@@ -1,0 +1,4 @@
+const getDayList = () => `<ul class="trip-days"></ul>`.trim();
+
+export {getDayList};
+
