+++ src/view/nav.js
@@ -1,0 +1,7 @@
+const getNav = () => `<nav class="trip-controls__trip-tabs  trip-tabs">
+              <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
+              <a class="trip-tabs__btn" href="#">Stats</a>
+            </nav>`.trim();
+
+export {getNav};
+
