+++ src/view/point-list.js
@@ -1,0 +1,4 @@
+const getPointList = () => `<ul class="trip-events__list"></ul>`.trim();
+
+export {getPointList};
+
