+++ src/view/total-cost.js
@@ -1,0 +1,5 @@
+const getTotalCost = () => `<p class="trip-info__cost">
+              Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
+            </p>`.trim();
+
+export {getTotalCost};
