+++ src/view/trip-info.js
@@ -1,0 +1,3 @@
+const getTripInfo = () => `<section class="trip-main__trip-info trip-info"></section>`.trim();
+
+export {getTripInfo};
